Improve sacct arguments
There are some subtleties regarding the interaction of sacct flags. The
command called now should only export details of jobs running between a
request start and end date, and give the elapsed time between those
dates. This is more suitable for determining usage from.

Comments explaining this have been added to the code but perhaps should
be discussed in a readme or documentation.